+++ .github/CHANGELOG.md
@@ -34,6 +34,7 @@ This is the changelog for the version V2.6.1. Don't forget to [join us on Discor
 * Bots use Squad-Perks (#101)
 * First implementation of Alternate Weapons on some vehicles (#172)
 * Spawn at Player, if player close to path
+* Don't use missiles on tdm / domination
 
 ### Bug fixes
 * #218 increased auto-sniper-time a little
@@ -45,6 +46,7 @@ This is the changelog for the version V2.6.1. Don't forget to [join us on Discor
 * several bugs in new node-editor
 
 * fix endless attack of vehicles
+* fix Gunmaster Gamemode
 
 ### New maps
 *